requirements.txt
# libraries for openai, chromadb, and python-dotenv
openai>=0.27.0
chromadb>=0.3.21
python-dotenv>=1.0.0